cloud_intro/end2end/requirements.txt
kafka-python==2.0.2
requests==2.32.3
psycopg2-binary==2.9.10
clickhouse-driver==0.2.9